+++ src/data/database.ts
@@ -205,6 +205,7 @@ export type Database = {
           id: string
           invited: boolean
           name: string
+          order: number
           venueid: string
         }
         Insert: {
@@ -214,6 +215,7 @@ export type Database = {
           id?: string
           invited: boolean
           name?: string
+          order?: number
           venueid: string
         }
         Update: {
@@ -223,6 +225,7 @@ export type Database = {
           id?: string
           invited?: boolean
           name?: string
+          order?: number
           venueid?: string
         }
         Relationships: [